requirements.txt
sseclient-py==1.8.0
python-dotenv==1.0.1
psutil==6.0.0